Removed: HTTP minitests (in favor of RSpec.)
