--- test/contrib/http/test_helper.rb
@@ -1,17 +1,0 @@
-require 'time'
-require 'net/http'
-require 'ddtrace'
-
-def wait_http_server(server, delay)
-  delay.times do |i|
-    uri = URI(server + '/')
-    begin
-      res = Net::HTTP.get_response(uri)
-      return true if res.code == '200'
-    rescue StandardError => e
-      puts e if i >= 3 # display errors only when failing repeatedly
-    end
-    sleep 1
-  end
-  false
-end
